feat: open the teaching guide from evaluations and seed Kat B texts
Folds the unused EvaluationModalNew into the main modal, lets staff change paid durations again, and fills empty driving-school templates.
